Dateien nach "source/usr/local/emhttp/plugins/lxc/include" hochladen

+++ source/usr/local/emhttp/plugins/lxc/include/Settings.php
@@ -6,12 +6,14 @@ class Settings {
   public $default_path;
   public $default_timeout;
   public $default_startdelay;
-  public $default_bridge;
-  public $available_bridges;
+  public $default_interface;
+  public $available_interfaces;
+  public $change_net_containers;
   public $status;
   public $default_cont_url;
   public $backup_enabled;
   public $backup_path;
+  public $backup_keep;
   public $backup_threads;
   public $backup_compression;
   public $backup_use_snapshot;
@@ -20,8 +22,8 @@ class Settings {
     $this->default_path = getVariable('/boot/config/plugins/lxc/lxc.conf', 'lxc.lxcpath');
     $this->default_timeout = getVariable('/boot/config/plugins/lxc/plugin.cfg', 'TIMEOUT');
     $this->default_startdelay = getVariable('/boot/config/plugins/lxc/plugin.cfg', 'AUTOSTART_DELAY');
-    $this->default_bridge = getVariable('/boot/config/plugins/lxc/default.conf', 'lxc.net.0.link');
-    $this->available_bridges = getAvailableBridges();
+    $this->default_interface = getVariable('/boot/config/plugins/lxc/default.conf', 'lxc.net.0.link');
+    $this->available_interfaces = getAvailableInterfaces();
     $this->status = getVariable('/boot/config/plugins/lxc/plugin.cfg', 'SERVICE');
     $this->default_cont_url = getVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_CONTAINER_URL');
     $this->backup_enabled = getVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_BACKUP_SERVICE');
@@ -32,8 +34,9 @@ class Settings {
     $this->backup_use_snapshot = getVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_BACKUP_USE_SNAPSHOT');
   }
 
-  function changeConfig($started, $default_path, $service, $timeout, $startdelay, $bridge, $default_cont_url, $backup_enabled, $backup_path, $backup_keep, $backup_threads, $backup_compression, $backup_use_snapshot) {
+  function changeConfig($started, $default_path, $service, $timeout, $startdelay, $interface, $change_net_containers, $default_cont_url, $backup_enabled, $backup_path, $backup_keep, $backup_threads, $backup_compression, $backup_use_snapshot) {
     $activeContainers = getActiveContainers();
+    $availContainers = getAllContainers();
 
     foreach ($activeContainers as $container) {
       exec('logger "LXC: Stopping container ' . $container . '"');
@@ -48,7 +51,6 @@ class Settings {
     setVariable('/boot/config/plugins/lxc/lxc.conf', 'lxc.lxcpath', $default_path);
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'TIMEOUT', $timeout);
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'AUTOSTART_DELAY', $startdelay);
-    setVariable('/boot/config/plugins/lxc/default.conf', 'lxc.net.0.link', $bridge);
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'SERVICE', $service);
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_CONTAINER_URL', $default_cont_url);
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_BACKUP_SERVICE', $backup_enabled);
@@ -58,7 +60,37 @@ class Settings {
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_BACKUP_COMPRESSION', $backup_compression);
     setVariable('/boot/config/plugins/lxc/plugin.cfg', 'LXC_BACKUP_USE_SNAPSHOT', $backup_use_snapshot);
 
-    unlink('/var/cache/lxc');
+    if (preg_match('/^(vhost|eth|bond)\d+$/', $interface)) {
+      exec('sed -i "/^lxc\.net\.0\.type/c lxc.net.0.type = macvlan" /boot/config/plugins/lxc/default.conf');
+      exec('sed -i "/^lxc\.net\.0\.link/c lxc.net.0.link = ' .$interface . '" /boot/config/plugins/lxc/default.conf');
+      if (strpos(file_get_contents('/boot/config/plugins/lxc/default.conf'), 'lxc.net.0.macvlan.mode = bridge') == false) {
+        exec('sed -i "/^lxc\.net\.0\.type/a lxc.net.0.macvlan.mode = bridge" /boot/config/plugins/lxc/default.conf');
+	    }
+    } else {
+      exec('sed -i "/^lxc\.net\.0\.type/c lxc.net.0.type = veth" /boot/config/plugins/lxc/default.conf');
+      exec('sed -i "/^lxc\.net\.0\.link/c lxc.net.0.link = ' . $interface . '" /boot/config/plugins/lxc/default.conf');
+      exec('sed -i "/^lxc\.net\.0\.macvlan\.mode/d" /boot/config/plugins/lxc/default.conf');
+    }
+
+    if ($change_net_containers == 'on' ) {
+      foreach ($availContainers as $container) {
+        if (preg_match('/^(vhost|eth|bond)\d+$/', $interface)) {
+          exec('sed -i "/^lxc\.net\.0\.type/c lxc.net.0.type = macvlan" ' . $default_path . '/' . $container->name . '/config');
+          exec('sed -i "/^lxc\.net\.0\.link/c lxc.net.0.link = ' . $interface . '" ' . $default_path . '/' . $container->name . '/config');
+          if (strpos(file_get_contents($default_path . '/' . $container->name . '/config'), 'lxc.net.0.macvlan.mode = bridge') == false) {
+            exec('sed -i "/^lxc\.net\.0\.type/a lxc.net.0.macvlan.mode = bridge" ' . $default_path . '/' . $container->name . '/config');
+          }
+        } else {
+          exec('sed -i "/^lxc\.net\.0\.type/c lxc.net.0.type = veth" ' . $default_path . '/' . $container->name . '/config');
+          exec('sed -i "/^lxc\.net\.0\.link/c lxc.net.0.link = ' . $interface . '" ' . $default_path . '/' . $container->name . '/config');
+          exec('sed -i "/^lxc\.net\.0\.macvlan\.mode/d" ' . $default_path . '/' . $container->name . '/config');
+        }
+      }
+    }
+  
+    if (is_dir('/var/cache/lxc')) {
+      unlink('/var/cache/lxc');
+    }
 
     if (!is_file('/etc/lxc/default.conf')) {
       symlink( "/boot/config/plugins/lxc/default.conf", "/etc/lxc/default.conf");
